Camera/src/main/java/com/umland/Configuration.java
package com.umland;

import com.fasterxml.jackson.databind.ObjectMapper;

import java.io.IOException;
import java.nio.charset.StandardCharsets;
import java.nio.file.Files;
import java.nio.file.Path;

public class Configuration {
    public static Configuration loadConfiguration(Path path) throws IOException {
        String content = new String(Files.readAllBytes(path), StandardCharsets.UTF_8);
        ObjectMapper objectMapper = new ObjectMapper();

        return objectMapper.readValue(content, Configuration.class);
    }

    private String factoryType;
    private String location;

    public String getFactoryType() {
        return factoryType;
    }

    public void setFactoryType(String factoryType) {
        this.factoryType = factoryType;
    }

    public String getLocation() {
        return location;
    }

    public void setLocation(String location) {
        this.location = location;
    }
}